update selectItemAddress
This restores the functionality of selecting items from the BDTH item
menue. There does exist an issue where if a new item is selected from
the plugin menue the original items position wil revert if the position
update was not finalized by manually de-selecting the original item by
clicking off it in-game.